+++ static/script/userprofile.js
@@ -1,85 +1,133 @@
-const domContainer = document.body;
-const root = ReactDOM.createRoot(domContainer);
+let ppanel = document.getElementById("posts_panel");
+let username = window.location.pathname.slice(window.location.pathname.lastIndexOf('/')+1);
 
-class API
+
+let user_x = '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="#6c757d" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-user-x"><path d="M16 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2"></path><circle cx="8.5" cy="7" r="4"></circle><line x1="18" y1="8" x2="23" y2="13"></line><line x1="23" y1="8" x2="18" y2="13"></line></svg>'
+let user_plus = '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="#0d6efd" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-user-plus"><path d="M16 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2"></path><circle cx="8.5" cy="7" r="4"></circle><line x1="20" y1="8" x2="20" y2="14"></line><line x1="23" y1="11" x2="17" y2="11"></line></svg>'
+let heart_0 = '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="#ff0000" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-heart"><path d="M20.84 4.61a5.5 5.5 0 0 0-7.78 0L12 5.67l-1.06-1.06a5.5 5.5 0 0 0-7.78 7.78l1.06 1.06L12 21.23l7.78-7.78 1.06-1.06a5.5 5.5 0 0 0 0-7.78z"></path></svg>';
+let heart_1 = '<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="#6c757d" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-heart"><path d="M20.84 4.61a5.5 5.5 0 0 0-7.78 0L12 5.67l-1.06-1.06a5.5 5.5 0 0 0-7.78 7.78l1.06 1.06L12 21.23l7.78-7.78 1.06-1.06a5.5 5.5 0 0 0 0-7.78z"></path></svg>';
+
+async function handleLike(event)
 {
-    async isAuthor()
+    console.log(event.currentTarget.textContent);
+    let pname = event.currentTarget.parentElement.parentElement.getAttribute('name');
+    let pnum = parseInt(pname.slice(pname.lastIndexOf('_')+1));
+    let is_liked = false;
+    if (event.currentTarget.classList.contains("liked"))
     {
-        let url = window.location.href.slice()
-        let data = {action: "isauthor"}
-        try {
-            // запрос
-            const response = await fetch(
-              url,
-              {
-                method: 'POST',
-                body: JSON.stringify(data),
+        event.currentTarget.classList.toggle("liked");
+        event.currentTarget.classList.toggle("notliked");
+        let count = parseInt(event.currentTarget.textContent);
+        count--;
+        event.currentTarget.innerHTML = heart_1 + count.toString();
+    }
+    else
+    {
+        is_liked = true;
+        event.currentTarget.classList.toggle("liked");
+        event.currentTarget.classList.toggle("notliked");
+        let count = parseInt(event.currentTarget.textContent);
+        count++;
+        event.currentTarget.innerHTML = heart_0 + count.toString();
+    }
+    try
+    {
+        url = window.location.href.slice();
+        const response = await fetch(
+            url,
+            {
+                method: "POST",
+                body: JSON.stringify({action: "like", postnum: pnum, isliked: is_liked}),
                 headers: {'Content-Type': 'application/json'}
-              }
-            );
-            // ответ
-            const json = await response.json();
-            console.log('isAuthor:', json["isauthor"]);
-            return json["isauthor"];
-          }
-          catch (error) {
-            console.error('Caught error:', error);
-          }
-    }
-}
-
-class TopPanel extends React.Component
-{
-    constructor(props)
-    {
-        super(props);
-        this.state = {
-            username: props["username"],
-            connection: new API()
-        }
-    }
-
-    async renderButtons()
-    {
-        let logout_icon = <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="white" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-log-out"><path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path><polyline points="16 17 21 12 16 7"></polyline><line x1="21" y1="12" x2="9" y2="12"></line></svg>;
-        let edit_icon = <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="white" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" className="feather feather-edit"><path d="M11 4H4a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7"></path><path d="M18.5 2.5a2.121 2.121 0 0 1 3 3L12 15l-4 1 1-4 9.5-9.5z"></path></svg>;
-        let ia = await this.state.connection.isAuthor();
-        console.log("renderButtons", ia)
-        if (ia)
-        {
-            return (
-                <div className="px-0">
-                    <button className="btn p-2 bg-transparent">{edit_icon}</button>
-                    <button className="btn p-2 bg-transparent">{logout_icon}</button>
-                </div>
-            )
-        }
-        else
-        {
-            return (
-                <div className="px-0">
-                    <button className="btn p-2 bg-transparent">{logout_icon}</button>
-                </div>
-            )
-        }
-    }
-
-    async render()
-    {
-        let buttons = await this.renderButtons();
-        return (
-            <div className="container-fluid d-flex bg-primary bg-gradient justify-content-between rounded-bottom px-0">
-                <div className="text-white d-flex align-items-center px-2">
-                    <h5 className="my-auto">{this.state.username}</h5>
-                </div>
-                {buttons}
-            </div>
+            }
         );
     }
+    catch (error)
+    {
+        console.error("Caught error:", error);
+    }
 }
 
-let username_read = window.location.pathname.slice(window.location.pathname.lastIndexOf("/")+1);
-let panel = new TopPanel({username: username_read})
-panel.render().then((elem) => {
-    root.render(elem);
-});
+function postHeader(author, datetime)
+{
+    let elem = document.createElement('div');
+    elem.className = "row justify-content-start d-flex align-items-center border-bottom border-secondary";
+    let ch1 = document.createElement('div');
+    ch1.style.width = 'fit-content';
+    ch1.innerHTML = `<h6>${author}</h6><small class="text-secondary">${datetime}</small>`;
+    elem.append(ch1);
+    return elem;
+}
+
+function postBody(text, imageUrl)
+{
+    let elem = document.createElement('div');
+    elem.className = "row";
+    let elem_text = document.createElement('p');
+    elem_text.innerText = text;
+    let elem_image = document.createElement('img');
+    elem_image.className='mx-auto';
+    elem_image.style.maxHeight = '500px';
+    elem_image.style.maxWidth = '500px';
+    elem_image.src = imageUrl;
+    elem.append(elem_text, elem_image);
+    return elem;
+}
+
+function postFooter(likes, liked)
+{
+    let elem = document.createElement('div');
+    elem.className = "row justify-content-between d-flex align-items-center border-top border-secondary";
+    let like_button = document.createElement('button');
+    if (liked)
+    {
+        like_button.className = "btn liked";
+        like_button.innerHTML = heart_0 + `${likes}`;
+    }
+    else
+    {
+        like_button.className = "btn notliked";
+        like_button.innerHTML = heart_1 + `${likes}`;
+    }
+    like_button.style.width = 'fit-content';
+    like_button.onclick = handleLike;
+    
+    elem.append(like_button);
+    return elem;
+}
+
+(async () => {
+    try
+    {
+        url = window.location.href.slice();
+        const response = await fetch(
+            url,
+            {
+                method: "POST",
+                body: JSON.stringify({action: "getposts", username: username}),
+                headers: {'Content-Type': 'application/json'}
+            }
+        );
+        const post_array = await response.json();
+        console.debug(post_array);
+
+        for (let i = 0; i < post_array.length; i++)
+        {
+            tpost = document.createElement('div');
+            tpost.setAttribute('name', `post_${post_array[i]["post_id"]}`)
+            tpost.className = "container-fluid border border-secondary rounded mt-2 bg-white";
+            tpost.style.height = 'fit-content';
+            tpost.append(
+                postHeader(post_array[i]["author"], post_array[i]["datetime"]),
+                postBody(post_array[i]["content"], window.location.origin + `/image/${post_array[i]["image_id"]}`),
+                postFooter(post_array[i]["likes"], post_array[i]["isliked"])
+            );
+            ppanel.append(tpost);
+        }
+
+    }
+    catch (error)
+    {
+        console.error("Caught error:", error)
+    }
+})();
