--- userprofile.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>{{ pagetitle }}</title>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-        <link href="{{ url_for('static', filename='style/profile.css') }}" rel="stylesheet" />
-    </head>
-    <body class="bg-transparent container mx-auto px-0" style="height: 100%;">
-        <!-- Top panel -->
-        <div class="container-fluid d-flex bg-primary bg-gradient justify-content-between rounded-bottom px-0">
-            <div class="text-white d-flex align-items-baseline px-2" style="width:fit-content;">
-                <h5 class="my-auto">{{ username }}</h5>
-            </div>
-            <div style="width: fit-content;">
-                {% if author %}
-                <a class="btn p-2" href="{{ url_for('user', username=username+'/new') }}"><i data-feather="edit" stroke="white"></i></a>
-                {% endif %}
-                <a class="btn p-2" href="/logout"><i data-feather="log-out" stroke="white"></i></a>
-            </div>
-        </div>
-        <!-- Posts panel -->
-        <div id="posts_panel" class="container-fluid d-flex flex-column align-items-start gap-1 bg-transparent px-0" style="height: 100%; overflow-y: scroll;">
-
-        </div>
-
-        <!-- <script src="https://unpkg.com/react@18/umd/react.development.js" crossorigin></script>
-        <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js" crossorigin></script>
-        <script src="https://unpkg.com/babel-standalone@6/babel.min.js"></script> -->
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/feather-icons/dist/feather.min.js"></script>
-        <script>feather.replace()</script>
-
-        <script src="{{ url_for('static', filename='script/userprofile.js') }}"></script>
-    </body>
-</html>
